requirements.txt
einops==0.8.1
numpy==1.21.5
pandas==1.4.4
Pillow==11.2.1
pytorch_msssim==1.0.0
torch==2.3.0+cu121
tqdm==4.64.1
